Fixes multiplying gamelist issue and adds in overlay and fast movement
Hopefully fixes issue #3.  Where the gamelist gets multiplied every time
the user exits a game.  Put a check in populateList() to return unless
the grid has less than the amount of files in System.

Adds in the Title Overlay.  This was already in place but I actually
just added it to GamelistGridView.cpp so now if you hold in a direction,
it will slowly pick up speed and when it hits a certain speed it will
overlay the first 2 letters of the game title.