Stability: slight modifications to Dynamic Loading
Dynamic loading will not update Cursor Range until it has finished
loading the prevous Range.  This could leave a user in a mix of unloaded
games but it keeps the system from blowing through the small amout of
vram it has.  (should stay below 80 loaded images)

Forces to unload ALL images if there are more than 80 loaded.

Loads Images faster to compensate for not loading on every move.

Slightly bigger loading range.

Now checks if the texture to be loaded actually exists

[ Needs to be cleaned up.  Some code went unused. ]