Allow Gridsize to be altered and more...
Gridsize can be changed in theme (rows and columns) by using:
<container name="gridRowsAndColumns"> <size>x y</size>

It does not auto size the tiles yet.
Not using this element will still draw the grid normally.

Allowed gridtile_background to be an Image as well as a Ninepatch
element.
This will allow for backwards compatability with upstream as its
Ninepatch themedata doesn't allow color.