Fix: Selected tile on bottom row disappearing
If the user moved down to a non-existant tile, the cursor would be
pushed back to an available tile.  This works fine but only 1 tile may
be selected at a time so the grid would be confused and always render
the non-existant tile.  It would use the tile from above in its place.

This is fixed by setting a tile's selected propery to false just before
it is set to hide.