Grid tiles now aligns flush with left/right sides
Cleaned up the god awful mess that built tile positions.  It is slightly
easier to read and will be easier to modify if need be in the future.

This will create a margin between tiles that is created by taking the
size of all x axies tiles, adding them up and subtracting their total to
the size of the grid.  This difference will become the new margin that
will be divided amongst all tiles (except tile 0)