es-core: Loop ComponentList by default.
This patch makes it possible to press up on the GuiMenu
to get to the last item, instead of pressing down multiple times.

This is especially usable when you want to shut down the system
frequently like on the Freeplay CM3.

Not sure why the less usable option here was the default setting,
since list looping is implemented.

With this patch shutting down the system takes 6 button presses
instead of 13.