Fix #457: Load delay during gamelist processing
The problem causing slow loading is because of how relative paths are handled. Both the game loader and the theme loader rely on the following functions: `resolveRelativePath`, `createRelativePath`, `removeCommonPath`.

These methods use `isDirectory()` to determine whether a file should search parent directories when referenced by `.` instead of the one it is in. This is relevant for themes which load parent relative content.

This is not relevant for loading or saving game lists which always use system root as the base path. In the old code every single file of every game list hits the file system multiple times to see if a path is a directory. If you're loading from the network or slow media this can take several minutes while every game is probed.

The fix adds `_searchParents` so that the theme system does (true) and the game loader does not (false).

You also need to set the `ParseGamelistOnly` setting to avoid touching the file system on load.

These changes bring an 8 minute load time down to about 3 seconds.

Added a comment as a warning to hopefully avoid re-introduction of `isDirectory()` in that path in the future.